patch IconContext out of phosphor-react
