+++ package.json
@@ -11,7 +11,8 @@
     "format": "npx prettier --write .",
     "format:check": "npx prettier --check .",
     "build-resume-pdf": "tsm ./scripts/buildResumePdf.mts",
-    "build-resume-txt": "tsm ./scripts/buildResumeTxt.mts"
+    "build-resume-txt": "tsm ./scripts/buildResumeTxt.mts",
+    "postinstall": "patch-package"
   },
   "dependencies": {
     "@sparticuz/chromium": "^108.0.1",
@@ -38,6 +39,7 @@
     "eslint-config-next": "^13.0.3",
     "eslint-config-prettier": "^8.5.0",
     "import-sort-style-atdrago": "^0.1.7",
+    "patch-package": "^6.5.0",
     "postcss": "^8.4.16",
     "prettier": "2.7.1",
     "prettier-plugin-import-sort": "^0.0.7",
